Created SQLite database and working search interface
I used SQLite to create a database containing 36 hikes. The UI has a
search option, with the ability to search by Name, Length or Difficulty.
After searching it displays the results in an Alert window. Would be
good to change this to showing the results on the page, below the search
option.