Updated Design of Search Results and Hike Info Pages
Updated the design of the search results so that each field (name,
length, etc.) of each hike would get its own line. With the hike
information pages, I added words for length and difficulty so that it
may be more understandable. For example, “Approximately 3 miles” for
length and “2 out of 4” for difficulty is displayed.